+++ R/model.R
@@ -32,8 +32,11 @@
 #' \dontrun{
 #' # One mutant:
 #' pfix_msa(ddg = 1.2, ddgact = 0.4, a1 = 1, a2 = 1)
+#'
 #' # A whole scan:
-#' spm <- generate_spm_data(znb_wt, seed = 1024)
+#' ex  <- function(f) system.file("extdata", f, package = "msamodel")
+#' wt  <- setup_enm(bio3d::read.pdb(ex("1znb_A.pdb")), node = "ca", d_max = 10.5)
+#' spm <- generate_spm_data(wt, seed = 1024)
 #' pfix_msa(spm$energy_data$ddg_jm, spm$energy_data$ddgact_jm, a1 = 1, a2 = 1)
 #' }
 #' @export
@@ -257,7 +260,10 @@ unimplemented_metric_message <- function(metric) {
 #' @family api
 #' @examples
 #' \dontrun{
-#' spm <- generate_spm_data(znb_wt, pdb_site_active = c(99,101,103,162,181,184,193,223), seed = 1024)
+#' ex  <- function(f) system.file("extdata", f, package = "msamodel")
+#' wt  <- setup_enm(bio3d::read.pdb(ex("1znb_A.pdb")), node = "ca", d_max = 10.5)
+#' act <- readr::read_csv(ex("znb_active_site.csv"))
+#' spm <- generate_spm_data(wt, pdb_site_active = act$pdb_site, seed = 1024)
 #' calculate_profiles(spm, a1 = 1, a2 = 1, metric = "nlrmsd")$site
 #' }
 #' @export
@@ -315,7 +321,10 @@ calculate_profiles <- function(spm, a1, a2, metric = c("lrmsd", "nlrmsd")) {
 #' @family api
 #' @examples
 #' \dontrun{
-#' spm <- generate_spm_data(znb_wt, pdb_site_active = c(99,101,103,162,181,184,193,223), seed = 1024)
+#' ex  <- function(f) system.file("extdata", f, package = "msamodel")
+#' wt  <- setup_enm(bio3d::read.pdb(ex("1znb_A.pdb")), node = "ca", d_max = 10.5)
+#' act <- readr::read_csv(ex("znb_active_site.csv"))
+#' spm <- generate_spm_data(wt, pdb_site_active = act$pdb_site, seed = 1024)
 #' calculate_decomposition(spm, a1 = 1, a2 = 1, metric = "nlrmsd")$site
 #' }
 #' @export